Simplify logic
